Corrections
Now Galileo works properly. Added check of the binary number "status"  to be sure that all the satellites implied in the position computation have the state "TOW Decoded" and the Galileo ones have as well the "EC1 2nd Code" status on. Implemented the 100ms mod necessary for Galileo with status E1C 2nd Code to work. Also changes getting lengths (changed to size) that could cause errors when there was less than 3 SVs.